refactor(ai-v4): harden browser-safe split build
Make the main build own cleanup, append the isolated Node-only test output,
and finalize declarations only after both builds succeed. Share common tsdown
options so the split configurations cannot drift.

Make the browser regression test bundle source instead of stale dist output,
always close Rspack, and execute in a VM without Node-only globals.
