+++ client-sdks/ai-sdk/src/__tests__/browser-bundle.test.ts
@@ -1,12 +1,12 @@
-import { execFile } from 'node:child_process';
-import { mkdtemp, rm, writeFile } from 'node:fs/promises';
+import { mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
-import { promisify } from 'node:util';
+import { TransformStream } from 'node:stream/web';
+import { URL } from 'node:url';
+import { TextDecoder, TextEncoder } from 'node:util';
+import { runInNewContext } from 'node:vm';
 import { rspack, type Stats } from '@rspack/core';
 import { afterAll, beforeAll, expect, it } from 'vitest';
-
-const execFileAsync = promisify(execFile);
 const packageRoot = path.resolve(import.meta.dirname, '../..');
 let tempDir: string;
 
@@ -23,7 +23,7 @@ it('loads toAISdkMessages in a browser bundle', async () => {
   const outputPath = path.join(tempDir, 'dist');
   await writeFile(
     entry,
-    `import { toAISdkMessages } from ${JSON.stringify(path.join(packageRoot, 'dist/ui.js'))};\n` +
+    `import { toAISdkMessages } from ${JSON.stringify(path.join(packageRoot, 'src/ui.ts'))};\n` +
       `globalThis.convertedMessages = toAISdkMessages(['hello']);\n`,
   );
 
@@ -32,28 +32,49 @@ it('loads toAISdkMessages in a browser bundle', async () => {
     target: 'web',
     entry,
     output: { path: outputPath, filename: 'bundle.js', publicPath: '' },
+    resolve: { extensions: ['.ts', '.js'] },
+    module: {
+      rules: [
+        {
+          test: /\.ts$/,
+          use: [{ loader: 'builtin:swc-loader', options: { jsc: { parser: { syntax: 'typescript' } } } }],
+        },
+      ],
+    },
     // No resolve.fallback: any Node built-in leaking into the browser graph
     // must fail the bundle step instead of being polyfilled away.
   });
 
-  const stats = await new Promise<Stats>((resolve, reject) => {
-    compiler.run((error, result) => {
-      if (error) reject(error);
-      else if (!result) reject(new Error('Rspack did not return build stats'));
-      else resolve(result);
-    });
-  });
-  await new Promise<void>((resolve, reject) => {
-    compiler.close(error => (error ? reject(error) : resolve()));
-  });
+  const stats = await (async () => {
+    try {
+      return await new Promise<Stats>((resolve, reject) => {
+        compiler.run((error, result) => {
+          if (error) reject(error);
+          else if (!result) reject(new Error('Rspack did not return build stats'));
+          else resolve(result);
+        });
+      });
+    } finally {
+      await new Promise<void>((resolve, reject) => {
+        compiler.close(error => (error ? reject(error) : resolve()));
+      });
+    }
+  })();
 
   expect(stats.hasErrors(), stats.toString({ errors: true })).toBe(false);
 
-  // Execute the browser bundle in Node; rspack dev bundles reference `self`,
-  // so alias it to globalThis first.
-  const { stdout } = await execFileAsync(process.execPath, [
-    '-e',
-    `globalThis.self = globalThis; require(${JSON.stringify(path.join(outputPath, 'bundle.js'))}); process.stdout.write(JSON.stringify(globalThis.convertedMessages));`,
-  ]);
-  expect(JSON.parse(stdout)).toMatchObject([{ role: 'user', parts: [{ type: 'text', text: 'hello' }] }]);
+  // Supply only the web globals touched during module initialization. Node-only
+  // globals such as require, process, and Buffer remain unavailable.
+  const sandbox: {
+    self?: unknown;
+    convertedMessages?: unknown;
+    TransformStream: typeof TransformStream;
+    TextDecoder: typeof TextDecoder;
+    TextEncoder: typeof TextEncoder;
+    URL: typeof URL;
+  } = { TransformStream, TextDecoder, TextEncoder, URL };
+  sandbox.self = sandbox;
+  runInNewContext(await readFile(path.join(outputPath, 'bundle.js'), 'utf8'), sandbox);
+
+  expect(sandbox.convertedMessages).toMatchObject([{ role: 'user', parts: [{ type: 'text', text: 'hello' }] }]);
 }, 30_000);
