+++ packages/_vendored/ai_v4/scripts/finalize-build.ts
@@ -1,0 +1,3 @@
+import { processAIDtsFiles } from './process-ai-dts-files.ts';
+
+await processAIDtsFiles();
